Fix Game Sound Effects with Settings Menu

+++ Assets/Scripts/PlayerSaveData.cs
@@ -11,8 +11,8 @@ public class PlayerSaveData : ScriptableObject
     public int health = 100;
     public int points = 0;
     public bool isLoadGame = false;
-    public int musicVolume = 100;
-    public int sfxVolume = 100;
+    public float musicVolume = 100;
+    public float sfxVolume = 100;
     public string previousScene = "MainMenu";
     public bool isSettingMenuOpened = false;
 }
